requirements.txt
absl-py==1.1.0
accelerate==0.24.1
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
bleach==6.1.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==3.0.0
cmake==3.27.7
cubinlinker-cu11==0.3.0.post1
cuda-python==11.8.3
cudf-cu11==23.10.0
cuml-cu11==23.10.0
cupy-cuda11x==12.2.0
cycler==0.11.0
dask==2023.9.2
dask-cuda==23.10.0
dask-cudf-cu11==23.10.0
Deprecated==1.2.13
diffusers==0.21.4
distributed==2023.9.2
docker-pycreds==0.4.0
einops==0.4.1
einops-exts==0.0.3
ema-pytorch==0.0.8
fastremap==1.14.1
fastrlock==0.8.2
filelock==3.9.0
fill-voids==2.0.6
fire==0.5.0
fonttools==4.34.4
frozenlist==1.3.0
fsspec==2023.10.0
ftfy==6.1.1
future==0.18.2
fvcore==0.1.5.post20221221
gdcm==1.1
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.26.2
google-auth-oauthlib==1.2.0
greenlet==3.0.1
grpcio==1.60.0
h5py==3.7.0
huggingface-hub==0.17.3
humanize==4.2.2
hydra-core==1.2.0
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
importlib-metadata==6.8.0
importlib-resources==5.9.0
iopath==0.1.10
Jinja2==3.1.2
joblib==1.1.0
kaggle==1.6.2
kiwisolver==1.4.3
lightning-utilities==0.10.0
lit==17.0.4
llvmlite==0.40.1
locket==1.0.0
lora-diffusion @ git+https://github.com/cloneofsimo/lora.git@bdd51b04c49fa90a88919a19850ec3b4cf3c5ecd
lpips==0.1.4
lungmask==0.2.19
lxml==4.9.1
Markdown==3.3.7
MarkupSafe==2.1.2
matplotlib==3.5.2
mediapipe==0.8.10
monai==1.3.0
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.2
mypy-extensions==1.0.0
networkx==2.8.5
nibabel==4.0.1
nilearn==0.9.1
numba==0.57.1
numpy==1.24.0
nvtx==0.2.8
oauthlib==3.2.0
omegaconf==2.2.3
open-clip-torch==2.23.0
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==1.4.3
partd==1.4.1
pathtools==0.1.2
Pillow==9.1.1
portalocker==2.8.2
promise==2.3
protobuf==4.23.0
psutil==5.9.6
ptxcompiler-cu11==0.7.0.post1
pyarrow==12.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyDeprecate==0.3.1
pydicom==2.3.0
pylibjpeg==2.0.0
pylibjpeg-libjpeg==2.0.0
pylibraft-cu11==23.10.0
pylidc==0.2.3
pynvml==11.4.1
pyparsing==3.1.1
pyre-extensions==0.0.23
python-dateutil==2.8.2
python-slugify==8.0.1
pytorch-lightning==1.9.1
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==19.0.2
raft-dask-cu11==23.10.0
regex==2022.6.2
requests==2.28.0
requests-oauthlib==1.3.1
rmm-cu11==23.10.0
rotary-embedding-torch==0.1.5
rsa==4.8
safetensors==0.4.0
scikit-image==0.19.3
scikit-learn==1.1.2
scikit-video==1.1.11
scipy==1.8.1
seaborn==0.11.2
sentencepiece==0.1.99
sentry-sdk==1.7.2
setproctitle==1.2.3
shortuuid==1.0.9
SimpleITK==2.1.1.2
six==1.16.0
smmap==5.0.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.23
submitit==1.5.0
sympy==1.12
tabulate==0.9.0
tblib==3.0.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2022.8.3
timm==0.9.12
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
torch==2.0.0+cu117
torch-tb-profiler==0.4.0
torchaudio==2.1.0+cu118
torchio==0.18.80
torchmetrics==0.10.3
torchvision==0.15.0+cu117
tornado==6.3.3
tqdm==4.64.0
transformers==4.35.2
treelite==3.9.1
treelite-runtime==3.9.1
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.2.0
ucx-py-cu11==0.34.0
urllib3==1.26.9
vtk==9.3.0
wandb==0.16.2
wcwidth==0.2.9
webencodings==0.5.1
Werkzeug==2.1.2
wrapt==1.14.1
xformers==0.0.18
yacs==0.1.8
yarl==1.7.2
zict==3.0.0
zipp==3.8.0
